user dashboard now populated with scores and highlighted content.
LESSON LEARNED:
if there is an issue accessing information that looks like it should behave a particular way, check the model.
because of a missing square bracket around the ranking value in the model, the queries returned, appeared to be an array of objects like they were intended to be
but did not behave as such and the information was inaccessible even though in the console it appeared to have square brackets around it. Array tests came up false while objects test came up true.

So, there is a way to display information in the console but not make it accessible.